# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WikiMind Gateway
  description: Local LLM Knowledge OS — Personal API
  version: 0.1.0
paths:
...
  /api/query:
    post:
      tags:
      - Query
      summary: Ask
      description: 'Ask a question against the wiki and receive an answer with citations.


        If request.conversation_id is None, a new conversation is created.

        Otherwise the question is appended as a new turn in the existing

        conversation.'
      operationId: ask_api_query_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AskResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/query/stream:
    post:
      tags:
      - Query
      summary: Ask Stream
      description: 'Stream an answer token-by-token via Server-Sent Events.


        Returns SSE events: ``chunk`` (text deltas), ``done`` (final AskResponse),

        or ``error``. The Query row is persisted only after the stream completes

        successfully. Client disconnect aborts without persisting.'
      operationId: ask_stream_api_query_stream_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/query/history:
    get:
      tags:
      - Query
      summary: Query History
      description: List past queries (legacy endpoint — UI uses /conversations instead).
      operationId: query_history_api_query_history_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/query/conversations:
    get:
      tags:
      - Query
      summary: List Conversations
      description: List conversations ordered by most recently updated first.
      operationId: list_conversations_api_query_conversations_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConversationSummary'
                title: Response List Conversations Api Query Conversations Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/query/conversations/{conversation_id}:
    get:
      tags:
      - Query
      summary: Get Conversation
      description: Return a single conversation with all its turns.
      operationId: get_conversation_api_query_conversations__conversation_id__get
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: string
          title: Conversation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/query/conversations/{conversation_id}/export:
    get:
      tags:
      - Query
      summary: Export Conversation
      description: Export a conversation as standalone markdown. Pure read, no DB
        writes.
      operationId: export_conversation_api_query_conversations__conversation_id__export_get
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: string
          title: Conversation Id
      responses:
        '200':
          description: Successful Response
          content:
            text/markdown:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/query/conversations/file-back:
    post:
      tags:
      - Query
      summary: File Back Selection
      description: File selected turns from one or more conversations back to the
        wiki as a single article.
      operationId: file_back_selection_api_query_conversations_file_back_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileBackSelectionRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/query/conversations/{conversation_id}/file-back:
    post:
      tags:
      - Query
      summary: File Back Conversation
      description: File the entire conversation back to the wiki as a single article.
      operationId: file_back_conversation_api_query_conversations__conversation_id__file_back_post
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: string
          title: Conversation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/query/conversations/{conversation_id}/crystallize:
    post:
      tags:
      - Query
      summary: Crystallize
      description: Distill a conversation into a new wiki article with page_type synthesis.
      operationId: crystallize_api_query_conversations__conversation_id__crystallize_post
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: string
          title: Conversation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrystallizeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/query/conversations/{conversation_id}/fork:
    post:
      tags:
      - Query
      summary: Fork Conversation
      description: 'Fork a conversation at a specific turn with a new question.


        Creates a new conversation that shares turns 0..turn_index-1 with the

        parent by reference. The original branch is preserved immutably.'
      operationId: fork_conversation_api_query_conversations__conversation_id__fork_post
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: string
          title: Conversation Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForkRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AskResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AskResponse:
      properties:
        query:
          $ref: '#/components/schemas/QueryResponse'
        conversation:
          $ref: '#/components/schemas/ConversationResponse'
      type: object
      required:
      - query
      - conversation
      title: AskResponse
      description: Response shape for POST /query — wraps both the new query and its
        parent conversation.
...
    CitationArticleRef:
      properties:
        slug:
          type: string
          title: Slug
        title:
          type: string
          title: Title
      type: object
      required:
      - slug
      - title
      title: CitationArticleRef
      description: Minimal reference to an article used inside a :class:`CitationResponse`.
    CitationResponse:
      properties:
        article:
          $ref: '#/components/schemas/CitationArticleRef'
        sources:
          items:
            $ref: '#/components/schemas/SourceResponse'
          type: array
          title: Sources
          default: []
        confidence_score:
          type: number
          title: Confidence Score
          default: 0.5
        effective_confidence:
          type: number
          title: Effective Confidence
          default: 0.5
      type: object
      required:
      - article
      title: CitationResponse
      description: 'A single Q&A citation: an article plus the sources it was compiled
        from.'
...
    ConversationDetail:
      properties:
        conversation:
          $ref: '#/components/schemas/ConversationResponse'
        queries:
          items:
            $ref: '#/components/schemas/QueryResponse'
          type: array
          title: Queries
      type: object
      required:
      - conversation
      - queries
      title: ConversationDetail
      description: Full conversation thread with all queries ordered by turn_index.
    ConversationResponse:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        filed_article_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Filed Article Id
        crystallized_article_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Crystallized Article Id
        parent_conversation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Parent Conversation Id
        forked_at_turn_index:
          anyOf:
          - type: integer
          - type: 'null'
          title: Forked At Turn Index
        fork_count:
          type: integer
          title: Fork Count
          default: 0
      type: object
      required:
      - id
      - title
      - created_at
      - updated_at
      title: ConversationResponse
      description: Conversation metadata exposed via API.
    ConversationSummary:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        filed_article_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Filed Article Id
        crystallized_article_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Crystallized Article Id
        parent_conversation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Parent Conversation Id
        forked_at_turn_index:
          anyOf:
          - type: integer
          - type: 'null'
          title: Forked At Turn Index
        fork_count:
          type: integer
          title: Fork Count
          default: 0
        turn_count:
          type: integer
          title: Turn Count
      type: object
      required:
      - id
      - title
      - created_at
      - updated_at
      - turn_count
      title: ConversationSummary
      description: Conversation summary for the history sidebar — adds turn count.
...
    CrystallizeResponse:
      properties:
        article_id:
          type: string
          title: Article Id
        article_slug:
          type: string
          title: Article Slug
        title:
          type: string
          title: Title
        turns_distilled:
          type: integer
          title: Turns Distilled
      type: object
      required:
      - article_id
      - article_slug
      - title
      - turns_distilled
      title: CrystallizeResponse
      description: Response after crystallizing a conversation into a wiki article.
...
    FileBackSelectionRequest:
      properties:
        selections:
          items:
            $ref: '#/components/schemas/TurnSelection'
          type: array
          title: Selections
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
      type: object
      required:
      - selections
      title: FileBackSelectionRequest
      description: Request to file back selected turns from one or more conversations.
    ForkRequest:
      properties:
        turn_index:
          type: integer
          title: Turn Index
        new_question:
          type: string
          title: New Question
      type: object
      required:
      - turn_index
      - new_question
      title: ForkRequest
      description: Request to fork a conversation at a specific turn with a new question.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    QueryRequest:
      properties:
        question:
          type: string
          maxLength: 10000
          title: Question
        file_back:
          type: boolean
          title: File Back
          default: false
        conversation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Conversation Id
      type: object
      required:
      - question
      title: QueryRequest
      description: Request to query the wiki.
    QueryResponse:
      properties:
        id:
          type: string
          title: Id
        question:
          type: string
          title: Question
        answer:
          type: string
          title: Answer
        confidence:
          anyOf:
          - type: string
          - type: 'null'
          title: Confidence
        source_article_ids:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Article Ids
        related_article_ids:
          anyOf:
          - type: string
          - type: 'null'
          title: Related Article Ids
        filed_back:
          type: boolean
          title: Filed Back
        filed_article_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Filed Article Id
        created_at:
          type: string
          format: date-time
          title: Created At
        conversation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Conversation Id
        turn_index:
          type: integer
          title: Turn Index
          default: 0
        citations:
          items:
            $ref: '#/components/schemas/CitationResponse'
          type: array
          title: Citations
          default: []
        wiki_worthiness:
          anyOf:
          - $ref: '#/components/schemas/WikiWorthinessScore'
          - type: 'null'
      type: object
      required:
      - id
      - question
      - answer
      - confidence
      - source_article_ids
      - related_article_ids
      - filed_back
      - filed_article_id
      - created_at
      title: QueryResponse
      description: 'Q&A response enriched with a full Answer → Article → Source citation
        chain.


        Mirrors the persisted :class:`Query` record fields while adding a

        resolved ``citations`` list so clients can see which articles were

        used and which original sources those articles came from.'
...
    SourceResponse:
      properties:
        id:
          type: string
          title: Id
        source_type:
          $ref: '#/components/schemas/SourceType'
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        source_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Url
        ingested_at:
          type: string
          format: date-time
          title: Ingested At
      type: object
      required:
      - id
      - source_type
      - title
      - source_url
      - ingested_at
      title: SourceResponse
      description: 'Provenance view of a raw ingested source exposed via the API.


        Trimmed view of :class:`Source` suitable for embedding in article and

        Q&A responses so callers can trace claims back to their origin

        (URL, PDF filename, upload date, etc.).'
    SourceType:
      type: string
      enum:
      - url
      - pdf
      - youtube
      - audio
      - text
      - rss
      - email
      - obsidian
      title: SourceType
      description: Type of ingested source.
...
    TurnSelection:
      properties:
        conversation_id:
          type: string
          title: Conversation Id
        turn_indices:
          items:
            type: integer
          type: array
          title: Turn Indices
      type: object
      required:
      - conversation_id
      - turn_indices
      title: TurnSelection
      description: A selection of specific turns from a single conversation.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WikiWorthinessScore:
      properties:
        word_count:
          type: integer
          title: Word Count
        source_count:
          type: integer
          title: Source Count
        synthesizes:
          type: boolean
          title: Synthesizes
        dedup_collision:
          type: boolean
          title: Dedup Collision
        passed:
          type: boolean
          title: Passed
        auto_filed:
          type: boolean
          title: Auto Filed
          default: false
      type: object
      required:
      - word_count
      - source_count
      - synthesizes
      - dedup_collision
      - passed
      title: WikiWorthinessScore
      description: 'Score describing whether a Q&A answer is worth filing back as
        a wiki page.


        Produced by the Q&A agent''s auto file-back scorer. ``passed`` is the

        overall verdict; ``auto_filed`` records whether a wiki article was

        actually created as a result of this score.'
...